arm: tegra: stingray: CONFIG_SWAP, CFQ scheduler as default (again)
I've got some upcoming plans for swapspace (namely, putting 1Gs worth
on a partition on the external SD-Card, as it's faster on writes than
our internal EMMC (believe it or not ...). Also, with the changes to
VM done a few weeks back in init.stingray.rc CFQ seems to now have an
advantage over the Deadline I/O scheduler (esp. as CFQ has had support
for handling Metadata writes specially on EMMC devices).

Change-Id: I7a0d84a07c41088c7f9aa3bc56431ca5cf434861